+++ src/rules/selector-combinator-space-before/__tests__/index.js
@@ -222,3 +222,20 @@ testRule(rule, {
     column: 16,
   } ],
 })
+
+testRule(rule, {
+  ruleName,
+  syntax: "less",
+  config: ["always"],
+
+  accept:[{
+    code: ".a when (@size>=60) and (@size<102) {}",
+    description: "ignore constructs",
+  }],
+
+  reject: [{
+    code: "a  +a {}",
+    description: "two spaces before + combinator",
+    message: messages.expectedBefore("+"),
+  }],
+})